Silence warnings about unused return values in ic
The warnings are:

oe_ei_encode_string.c:26:3: warning: expression result unused [-Wunused-value]
  (int) ei_encode_string(0,&size,p);
  ^     ~~~~~~~~~~~~~~~~~~~~~~~~~~~

oe_ei_encode_port.c:26:3: warning: expression result unused [-Wunused-value]
  (int) ei_encode_port(NULL, &size, p);
  ^     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

oe_ei_encode_atom.c:26:3: warning: expression result unused [-Wunused-value]
  (int) ei_encode_atom(0,&size,p);
  ^     ~~~~~~~~~~~~~~~~~~~~~~~~~

oe_ei_encode_pid.c:26::3: warning: expression result unused [-Wunused-value]
  (int) ei_encode_pid(NULL, &size, p);
  ^     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

oe_ei_encode_ref.c:26:3: warning: expression result unused [-Wunused-value]
  (int) ei_encode_ref(NULL, &size, p);
  ^     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

oe_ei_encode_term.c:26:3: warning: expression result unused [-Wunused-value]
  (int) ei_encode_term(NULL, &size, t);
  ^     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ lib/ic/c_src/oe_ei_encode_term.c
@@ -23,7 +23,7 @@
 int oe_ei_encode_term(CORBA_Environment *ev, void *t) {
   int size = ev->_iout;
   
-  (int) ei_encode_term(NULL, &size, t);
+  ei_encode_term(NULL, &size, t);
   
   if (size >= ev->_outbufsz) {
     char *buf = ev->_outbuf;
